chore: Add UI, tsc and eslint packages
