+++ web_ui/packages/eslint-config/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@geti/eslint-config",
+  "type": "commonjs",
+  "version": "1.0.0",
+  "main": "eslint.config.mjs",
+  "devDependencies": {
+  },
+  "peerDependencies": {
+    "eslint": "^9.25.1",
+    "@eslint/compat": "^1.2.8",
+    "@typescript-eslint/eslint-plugin": "^8.31.1",
+    "eslint-plugin-header": "^3.1.1",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jest": "^28.11.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "description": ""
+}
