+++ web_ui/packages/typescript-config/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@geti/typescript-config",
+  "version": "1.0.0",
+  "main": "tsconfig.base.json",
+  "devDependencies": {},
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "peerDependencies": {
+    "typescript": "^5.7.2"
+  },
+  "author": "",
+  "license": "ISC",
+  "description": ""
+}
