+++ web_ui/packages/ui/eslint.config.mjs
@@ -1,0 +1,3 @@
+import sharedConfig from '@geti/eslint-config';
+
+export default sharedConfig;
