+++ web_ui/packages/ui/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@geti/ui",
+  "version": "1.0.0",
+  "description": "Contains UI representative components",
+  "main": "index.js",
+  "devDependencies": {
+    "@geti/typescript-config": "*",
+    "@geti/eslint-config": "*",
+    "eslint": "^9.25.1",
+    "typescript": "^5.7.2"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "type-check": "tsc --noEmit -p tsconfig.json",
+    "lint": "eslint --config eslint.config.mjs \"src/**/*.ts*\" --max-warnings 0",
+    "lint:fix": "eslint --fix --config eslint.config.mjs \"./src/**/*.ts*\""
+  },
+  "author": "",
+  "license": "ISC"
+}
